dk3v1/requirements.txt
PyMuPDF
pdfplumber
Pillow
numpy
torch
opencv-python
groq
